# Push both gRPC and Connect RPC attributes