Resolve #186

Changes:
- Push grpc and connect rpc attributes at the same time.
- Drop grpc web attributes.

The result of `benchstat`

```sh
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: connectrpc.com/otelconnect
cpu: Apple M1 Max
                            │   old.txt   │              new.txt               │
                            │   sec/op    │   sec/op     vs base               │
StreamingBase-10              71.91µ ± 6%   71.19µ ± 8%       ~ (p=0.631 n=10)
StreamingWithInterceptor-10   73.40µ ± 2%   73.22µ ± 1%       ~ (p=0.579 n=10)
UnaryBase-10                  45.37µ ± 3%   45.04µ ± 4%       ~ (p=0.579 n=10)
UnaryWithInterceptor-10       45.86µ ± 1%   45.83µ ± 1%       ~ (p=0.796 n=10)
geomean                       57.57µ        57.27µ       -0.51%

                            │    old.txt    │               new.txt               │
                            │     B/op      │     B/op      vs base               │
StreamingBase-10              24.82Ki ± 10%   26.58Ki ± 6%       ~ (p=0.165 n=10)
StreamingWithInterceptor-10   31.35Ki ±  6%   31.59Ki ± 4%       ~ (p=0.280 n=10)
UnaryBase-10                  19.85Ki ±  7%   19.52Ki ± 5%       ~ (p=0.247 n=10)
UnaryWithInterceptor-10       24.07Ki ±  2%   24.29Ki ± 4%       ~ (p=0.393 n=10)
geomean                       24.69Ki         25.12Ki       +1.73%

                            │  old.txt   │               new.txt               │
                            │ allocs/op  │ allocs/op   vs base                 │
StreamingBase-10              172.0 ± 0%   172.0 ± 0%       ~ (p=1.000 n=10) ¹
StreamingWithInterceptor-10   233.0 ± 0%   233.0 ± 0%       ~ (p=1.000 n=10) ¹
UnaryBase-10                  152.0 ± 0%   152.0 ± 0%       ~ (p=1.000 n=10) ¹
UnaryWithInterceptor-10       202.0 ± 0%   202.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                       187.3        187.3       +0.00%
¹ all samples are equal
```