@jhump 2025-05-19
@XSAM, let's try to reach consensus in the issue (#186) before proceeding further with this PR.